New lore trait and suggested cleanup
Also includes fixing of typos and the data not being included in item tasks

+++ src/main/kotlin/world/cepi/itemextension/item/traits/list/AttackSpeedTrait.kt
@@ -5,7 +5,7 @@ import world.cepi.itemextension.command.plus
 import world.cepi.itemextension.item.traits.Trait
 
 class AttackSpeedTrait(
-        val attackSpeed: Int
+        val attackSpeed: Double
 ) : Trait {
 
     override val loreIndex = 4
